små fisk
meget små ting der blev fikset

+++ Library/File.cs
@@ -44,7 +44,7 @@ namespace Library
             SubFiles = new List<FileVersion>();
         }
 
-        public File(int id, string title, string description, double verionNr, User fileLock, DateTime fileLockTime)
+        public File(int id, string title, string description, int verionNr, User fileLock, DateTime fileLockTime)
         {
             Id = id;
             Title = title;
